+++ components/TeachersTable.tsx
@@ -1,20 +1,14 @@
 import { Separator } from './ui/separator';
 import { getAvatarName } from '@/lib/utils';
-import { cached_teachers, uncached_teachers } from '@/prisma/db-calls';
 import Link from 'next/link';
 import Section from './Section';
 import { Avatar, AvatarFallback } from './ui/avatar';
-import { getServerSession } from 'next-auth';
-import { authOptions } from '@/lib/auth';
-import { redirect } from 'next/navigation';
+import { cached_teachers } from '@/actions';
 
 export const columns = ['Avatar', 'Name', 'Subject'];
 
 async function TeachersTable() {
-  const session = await getServerSession(authOptions);
-  if (!session) redirect(`/api/auth/signin`);
-
-  const teachers = await cached_teachers(session.user.id);
+  const teachers = await cached_teachers();
 
   if (!teachers || !teachers.length)
     return (
